# swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Values Api"},"paths":{"/api/AnimalCategory":{"get":{"tags":["AnimalCategory"],"operationId":"GetAll","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/AnimalCategoryViewModel"}}}}}},"/api/Order/{id}":{"get":{"tags":["Order"],"operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/OrderViewModel"}}}},"put":{"tags":["Order"],"operationId":"Update","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"order","in":"body","required":false,"schema":{"$ref":"#/definitions/OrderViewModel"}},{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["Order"],"operationId":"Delete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/Order":{"post":{"tags":["Order"],"operationId":"Add","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"order","in":"body","required":false,"schema":{"$ref":"#/definitions/OrderViewModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Product/{id}":{"get":{"tags":["Product"],"operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/ProductViewModel"}}}}},"/api/ShortProduct/animalCategoryId":{"get":{"tags":["ShortProduct"],"operationId":"GetByCategory","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"animalCategoryId","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/ShortProductViewModel"}}}}}}},"definitions":{"AnimalCategoryViewModel":{"type":"object","properties":{"name":{"type":"string"},"id":{"format":"int32","type":"integer"}}},"OrderViewModel":{"type":"object","properties":{"totalPrice":{"format":"double","type":"number"},"productCount":{"format":"int32","type":"integer"},"productId":{"format":"int32","type":"integer"},"product":{"$ref":"#/definitions/ProductViewModel"},"timeStamp":{"format":"date-time","type":"string"},"id":{"format":"int32","type":"integer"}}},"ProductViewModel":{"type":"object","properties":{"description":{"type":"string"},"animalCategoryId":{"format":"int32","type":"integer"},"animalCategory":{"$ref":"#/definitions/AnimalCategoryViewModel"},"name":{"type":"string"},"price":{"format":"double","type":"number"},"id":{"format":"int32","type":"integer"}}},"ShortProductViewModel":{"type":"object","properties":{"name":{"type":"string"},"price":{"format":"double","type":"number"},"id":{"format":"int32","type":"integer"}}}},"securityDefinitions":{"Bearer":{"name":"Authorization","in":"header","type":"apiKey"}},"security":[{"Bearer":[]}]}